#60 Forms is still work in progress
Form sets and form sections can be added or deleted based on a mocked
API that stores data in json files.
Questions can be only viewed and the edit part is the work in progress.
To avoid the existing functionality for answers and emergencies, I chose
to create a new model, namely EditableForms. Once the edit part is done,
I plan to merge the existing Forms model int oa single one.